Move command loader functions to utilities file
