+++ lib/cmds/reload.js
@@ -7,6 +7,7 @@ exports.info = {
 
 const config = require('../../config');
 const logger = require('winston');
+const utils = require('../utilities');
 //const twitter = require('../twitter')();
 
 exports.run = function Reload(client, msg, params = []) {
@@ -26,7 +27,7 @@ exports.run = function Reload(client, msg, params = []) {
     params.forEach(x => x.toLowerCase());
     if (params[0] === 'all') {
       logger.debug('reloading all commands');
-      client.loadAllCommands(client)
+      utils.loadAllCommands(client)
         .then(() => {
           msg.reply(`${client.commands.size} command${client.commands.size === 1 ? '' : 's'} ` +
             `have been loaded successfully.`);
@@ -36,14 +37,14 @@ exports.run = function Reload(client, msg, params = []) {
     }
 
     /*if (params[0] === 'twitterclient') {
-      msg.reply(`Twitter Client reset.`);
-      logger.debug('Manually resetting twitter client');
-      twitter.reset(client);
-      utils.finish(msg, exports.name);
-      return;
-    }*/
+     msg.reply(`Twitter Client reset.`);
+     logger.debug('Manually resetting twitter client');
+     twitter.reset(client);
+     utils.finish(msg, exports.name);
+     return;
+     }*/
 
-    client.loadOneCommand(client, params[0])
+    utils.loadOneCommand(client, params[0])
       .then(() => {
         msg.reply(`Command **${params[0]}** loaded successfully`);
         resolve();
