+++ lib/index.js
@@ -45,10 +45,8 @@ client.count = {
 
 // Instantiate other modules into the client object
 client.mongo = require('./mongo');
-client.loadOneCommand = require('./loadOneCommand');
-client.loadAllCommands = require('./loadAllCommands');
 
-client.loadAllCommands(client)
+utils.loadAllCommands(client)
   .then(() => {
     logger.info(`Loaded ${client.commands.size} command${client.commands.size === 1 ? '' : 's'}, ` +
       `with ${client.aliases.size} aliases.`);
