+++ lib/utilities.js
@@ -1,5 +1,6 @@
 'use strict';
 const logger = require('winston');
+const fs = require('fs');
 const path = require('path');
 const now = require('performance-now');
 const request = require('request');
@@ -49,5 +50,70 @@ module.exports = {
       });
     });
   },
+
+  loadOneCommand: (client, file) => {
+    return new Promise((resolve, reject) => {
+      // Get the file (cmd) name
+      const cmdPath = path.join(client.commandsDir, file);
+      const name = path.parse(cmdPath).name;
+      // See if the command exists already in our collection
+      if (client.commands.has(name)) {
+        // Command exists already
+        module.exports.removeCommand(client, name);
+      }
+      // Load the command
+      logger.debug(`Loading command: '${name}'`);
+      try {
+        const command = require(cmdPath);
+        // Set name from file name. Used in cmds command
+        command.name = name;
+        // Add command to Collection
+        client.commands.set(name, command);
+        // Add any command aliases to the alias Collection
+        if (command.info.aliases) {
+          command.info.aliases.forEach(alias => {
+            // First come first serve in case of duplicates
+            if (client.commands.has(alias)) {
+              logger.warn(`The command ${name} alias ${alias} already exists.`);
+            } else {
+              client.aliases.set(alias, command);
+            }
+          });
+        }
+        resolve();
+      } catch (e) {
+        reject(e);
+      }
+    });
+  },
+
+
+  loadAllCommands: (client) => {
+    return new Promise((resolve, reject) => {
+      logger.debug('Loading commands into memory');
+      fs.readdir(client.commandsDir, (err, files) => {
+        if (err) {
+          reject(err);
+        } else {
+          logger.info(`Loading ${files.length} command(s).`);
+          // Create promise array of promises to load the command files
+          const promiseArray = files.map(file => module.exports.loadOneCommand(client, file));
+          // Flush any commands that no longer have a file present
+          client.commands
+            .map(c => c.name)
+            .filter(c => files.indexOf(`${c}.js`) === -1)
+            .forEach(c => {
+              logger.debug(`Command file '${c}' not found.`);
+              module.exports.removeCommand(client, c);
+            });
+          // After no longer existing commands are flushed, load all commands
+          Promise.all(promiseArray)
+            .then(resolve)
+            .catch(reject);
+        }
+      });
+    });
+  },
+
 
 };
